Merge quick changes from web interface

+++ commands/deletereminder.js
@@ -5,11 +5,10 @@ module.exports = {
     args: true,
     usage: '<reminder-id>',
     async execute(message, args) {
-        const {Reminders, mysql} = require('../index');
+        const {mysql} = require('../index');
         const messageOwner = message.author.id;
         
-        //const rowCount = await Reminders.destroy({where: {reminder_id: args[0], username: messageOwner}});
-        mysql.query("DELETE FROM tbl_Reminders WHERE reminder_id = " + mysql.escape(args[0]), function (error, result) {
+        mysql.query("DELETE FROM tbl_Reminders WHERE reminder_id = " + mysql.escape(args[0]) + " AND username = " + mysql.escape(messageOwner), function (error, result) {
             if (error) throw error;
             if (result.affectedRows == 0) {
                 return message.reply('You don\'t have a reminder with that ID.');
@@ -19,10 +18,5 @@ module.exports = {
                 return message.reply('Reminder '+args[0]+' has been deleted!');
             }
         });
-
-        /*if (!rowCount) {
-            return message.reply('You don\'t have a reminder with that ID.');
-        }*/
-        //return message.reply('Reminder '+args[0]+' has been deleted!');
     },
 };
